feat(specificMentor): create mentor_controller file
- add mentor_controller

[Finishes #168411694]
