+++ server/routes/mentor.js
@@ -1,11 +1,11 @@
 import express from 'express';
-import UserController from '../controllers/user_controller';
+import MentorController from '../controllers/mentor_controller';
 import auth from '../middleware/auth';
 
 
 const router = express.Router();
 
-const { getAllMentors, getSpecificMentor } = UserController;
+const { getAllMentors, getSpecificMentor } = MentorController;
 
 router.get('/mentors', auth.auth, getAllMentors);
 router.get('/mentors/:id', auth.auth, getSpecificMentor);
